Refactor BinaryChoice component
